requirements.txt
streamlit==1.39.0
boto3==1.35.0
mcp==1.1.0
anthropic==0.40.0
langchain-aws==0.2.6
langchain-anthropic==0.2.4
langchain-core==0.3.21
langchain==0.3.7
python-dotenv==1.0.0
pandas==2.2.0
plotly==5.24.1
requests==2.31.0
botocore==1.35.0
